idk what Ive done honestly at this point
Im just waiting for the pizza to arrive and  they said it was 2 hours until they arrive, and its been 3, so Im pretty disappointed at this point because I didn't want to eat anything and now I'm just counting on them to deliver the pizza, so all in all, its a pretty annoying situation that I've been put in. I'd like to eat and then work on the project more because at this point I cant really think at all, but I dont know if this requires any thinking to do this as the code already is a bit of a spaghetti, but we'll see, I hope for the pizzas to arrive soon.